refactor(controller): decouples HttpValidator from controller
