+++ backend/src/main/modules/presentation/controllers/create-user.controller.ts
@@ -1,18 +1,19 @@
-import { HttpValidator } from "@presentation/validation/http-validator";
+import type { HttpValidator } from "@presentation/contracts/http-validator.contract";
 import type { CreateUser } from "../../domain/usecases/create-user";
 import type { HttpRequest } from "../dtos/http-request.dto";
 import type { HttpResponse } from "../dtos/http-response.dto";
 import { ErrorPresenter } from "../errors/error-presenter";
 import { created } from "../http/http-helpers";
-import { createUserSchema } from "@presentation/validation/create-user.schema";
 
 export class CreateUserController {
-  constructor(private readonly createUser: CreateUser) {}
+  constructor(
+    private readonly createUser: CreateUser,
+    private readonly httpValidator: HttpValidator,
+  ) {}
 
   async handle(request: HttpRequest): Promise<HttpResponse> {
     try {
-      const httpValidator = new HttpValidator(createUserSchema);
-      const userDTO = httpValidator.execute(request);
+      const userDTO = this.httpValidator.execute(request);
 
       const { name, email, password } = userDTO;
       const user = await this.createUser.execute({
