+++ backend/src/main/modules/presentation/validation/zod-http-validator.ts
@@ -1,10 +1,11 @@
 import { AppError } from "@presentation/errors/app-error";
 import type { CreateUserDTO } from "@domain/usecases/create-user";
 import type { HttpRequest } from "@presentation/dtos/http-request.dto";
-import type { createUserSchema } from "../validation/create-user.schema";
+import type { createUserSchema } from "./create-user.schema";
+import type { HttpValidator } from "@presentation/contracts/http-validator.contract";
 import type { z } from "zod";
 
-export class HttpValidator {
+export class ZodHttpValidator implements HttpValidator {
   constructor(private readonly schema: z.ZodType) {}
   execute(request: HttpRequest): CreateUserDTO {
     type Schema = z.infer<typeof createUserSchema>;
